# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sync/snapshot:
    get:
      operationId: getSyncSnapshot
      tags: [sync]
      summary: Pull the full per-user sync snapshot
      description: |
        Returns every per-user document the server has stored for the
        authenticated user (profile, daily nutrition / activity / meal-plan
        rows, wellbeing entries, gamification state, badges, assessments,
        supplements). Used on app launch to reconcile AsyncStorage to the
        latest server view (server wins on conflict — last-write-wins).
      responses:
        "200":
          description: Snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncSnapshot"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /sync/profile:
    put:
      operationId: putSyncProfile
      tags: [sync]
      summary: Replace the user's profile + preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncProfileBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sync/nutrition/{date}:
    put:
      operationId: putSyncNutritionDay
      tags: [sync]
      summary: Upsert one day's nutrition log
      parameters:
        - $ref: "#/components/parameters/DayParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncDayBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sync/activity/{date}:
    put:
      operationId: putSyncActivityDay
      tags: [sync]
      summary: Upsert one day's activity log
      parameters:
        - $ref: "#/components/parameters/DayParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncDayBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sync/meal-plan/{date}:
    put:
      operationId: putSyncMealPlanDay
      tags: [sync]
      summary: Upsert one day's meal plan + dietary prefs + favourites
      parameters:
        - $ref: "#/components/parameters/DayParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncDayBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sync/supplements:
    put:
      operationId: putSyncSupplements
      tags: [sync]
      summary: Replace the user's supplement state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncStateBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sync/gamification:
    put:
      operationId: putSyncGamification
      tags: [sync]
      summary: Replace the user's gamification state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncStateBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sync/wellbeing:
    post:
      operationId: postSyncWellbeingEntry
      tags: [sync]
      summary: Append a wellbeing / session entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncWellbeingEntryBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sync/assessment:
    post:
      operationId: postSyncAssessment
      tags: [sync]
      summary: Append a clinical assessment (DEXA / FRAX)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncAssessmentBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sync/outcomes:
    post:
      operationId: postSyncOutcome
      tags: [sync]
      summary: Append an idempotent structured self-reported outcome check-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncOutcomeBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  parameters:
    DayParam:
      name: date
      in: path
      required: true
      description: Local-time YYYY-MM-DD calendar day.
      schema:
        type: string
        pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SyncError"
    BadRequest:
      description: Body / params failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SyncError"
...
  schemas:
...
    SyncOk:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    SyncError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SyncProfileBody:
      type: object
      additionalProperties: false
      properties:
        profile:
          type: object
          additionalProperties: true
        updatedAtMs:
          type: integer
      required:
        - profile
...
    SyncDayBody:
      description: |
        Per-day domain document body. The `data` payload is opaque to the
        server (validated only as a JSON object with bounded size); the
        client is the source of truth for the inner shape so we don't need
        to migrate the API every time a new field is added.
      type: object
      additionalProperties: false
      properties:
        data:
          type: object
          additionalProperties: true
        updatedAtMs:
          type: integer
      required:
        - data
    SyncStateBody:
      description: |
        Full-state replace body. `state` is opaque to the server (bounded
        size, JSON object).
      type: object
      additionalProperties: false
      properties:
        state:
          type: object
          additionalProperties: true
        updatedAtMs:
          type: integer
      required:
        - state
    SyncWellbeingEntryBody:
      type: object
      additionalProperties: false
      properties:
        entryId:
          type: string
          minLength: 1
          maxLength: 200
        entry:
          type: object
          additionalProperties: true
        completedAtMs:
          type: integer
      required:
        - entryId
        - entry
        - completedAtMs
    SyncAssessmentBody:
      type: object
      additionalProperties: false
      properties:
        resultId:
          type: string
          minLength: 1
          maxLength: 200
        kind:
          type: string
          minLength: 1
          maxLength: 64
        payload:
          type: object
          additionalProperties: true
        takenAtMs:
          type: integer
      required:
        - resultId
        - kind
        - payload
        - takenAtMs
    SyncOutcomeBody:
      type: object
      additionalProperties: false
      properties:
        entryId:
          type: string
          minLength: 1
          maxLength: 100
        entry:
          type: object
          additionalProperties: true
        recordedAtMs:
          type: integer
      required: [entryId, entry, recordedAtMs]
    SyncOutcomeSnapshotEntry:
      type: object
      additionalProperties: false
      properties:
        entryId:
          type: string
        entry:
          type: object
          additionalProperties: true
        recordedAtMs:
          type: integer
      required: [entryId, entry, recordedAtMs]
    SyncProfileSnapshot:
      type: object
      additionalProperties: false
      properties:
        profile:
          type: object
          additionalProperties: true
        updatedAtMs:
          type: integer
          nullable: true
      required:
        - profile
    SyncDaySnapshot:
      type: object
      additionalProperties: false
      properties:
        day:
          type: string
        data:
          type: object
          additionalProperties: true
        updatedAtMs:
          type: integer
          nullable: true
      required:
        - day
        - data
    SyncStateSnapshot:
      type: object
      additionalProperties: false
      properties:
        state:
          type: object
          additionalProperties: true
        updatedAtMs:
          type: integer
          nullable: true
      required:
        - state
    SyncWellbeingSnapshotEntry:
      type: object
      additionalProperties: false
      properties:
        entryId:
          type: string
        entry:
          type: object
          additionalProperties: true
        completedAtMs:
          type: integer
      required:
        - entryId
        - entry
        - completedAtMs
    SyncAssessmentSnapshotEntry:
      type: object
      additionalProperties: false
      properties:
        resultId:
          type: string
        kind:
          type: string
        payload:
          type: object
          additionalProperties: true
        takenAtMs:
          type: integer
      required:
        - resultId
        - kind
        - payload
        - takenAtMs
    SyncSnapshot:
      type: object
      additionalProperties: false
      properties:
        appUserId:
          type: string
        profile:
          oneOf:
            - $ref: "#/components/schemas/SyncProfileSnapshot"
            - type: "null"
        nutrition:
          type: array
          items:
            $ref: "#/components/schemas/SyncDaySnapshot"
        activity:
          type: array
          items:
            $ref: "#/components/schemas/SyncDaySnapshot"
        mealPlan:
          type: array
          items:
            $ref: "#/components/schemas/SyncDaySnapshot"
        wellbeing:
          type: array
          items:
            $ref: "#/components/schemas/SyncWellbeingSnapshotEntry"
        gamification:
          oneOf:
            - $ref: "#/components/schemas/SyncStateSnapshot"
            - type: "null"
        supplements:
          oneOf:
            - $ref: "#/components/schemas/SyncStateSnapshot"
            - type: "null"
        assessments:
          type: array
          items:
            $ref: "#/components/schemas/SyncAssessmentSnapshotEntry"
        outcomes:
          type: array
          items:
            $ref: "#/components/schemas/SyncOutcomeSnapshotEntry"
      required:
        - appUserId
        - profile
        - nutrition
        - activity
        - mealPlan
        - wellbeing
        - gamification
        - supplements
        - assessments
        - outcomes
...