# lib/api-zod/src/generated/types/syncOutcomeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SyncOutcomeBodyEntry } from "./syncOutcomeBodyEntry";

export interface SyncOutcomeBody {
  /**
   * @minLength 1
   * @maxLength 100
   */
  entryId: string;
  entry: SyncOutcomeBodyEntry;
  recordedAtMs: number;
}