# lib/api-zod/src/generated/types/syncOutcomeSnapshotEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SyncOutcomeSnapshotEntryEntry } from "./syncOutcomeSnapshotEntryEntry";

export interface SyncOutcomeSnapshotEntry {
  entryId: string;
  entry: SyncOutcomeSnapshotEntryEntry;
  recordedAtMs: number;
}