# lib/api-zod/src/generated/types/syncSnapshot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SyncAssessmentSnapshotEntry } from "./syncAssessmentSnapshotEntry";
import type { SyncDaySnapshot } from "./syncDaySnapshot";
import type { SyncOutcomeSnapshotEntry } from "./syncOutcomeSnapshotEntry";
import type { SyncProfileSnapshot } from "./syncProfileSnapshot";
import type { SyncStateSnapshot } from "./syncStateSnapshot";
import type { SyncWellbeingSnapshotEntry } from "./syncWellbeingSnapshotEntry";

export interface SyncSnapshot {
  appUserId: string;
  profile: SyncProfileSnapshot | null;
  nutrition: SyncDaySnapshot[];
  activity: SyncDaySnapshot[];
  mealPlan: SyncDaySnapshot[];
  wellbeing: SyncWellbeingSnapshotEntry[];
  gamification: SyncStateSnapshot | null;
  supplements: SyncStateSnapshot | null;
  assessments: SyncAssessmentSnapshotEntry[];
  outcomes: SyncOutcomeSnapshotEntry[];
}